Fixed crash from TaskbarProgress when BuiltInComInteropSupport is disabled. (#2255)
Fixed terminal sequence printed to console in legacy console.
Include progress in error and warning states.
